Feat move warehouse to admin layout (#40)
* remove old warehouse page

* add warehouse to admin

* add tooltip to create button

---------
